test(e2e): speed up tests by moving them here
